Improve handling EOF and EOS conditions
On transmit, send EOT pseudo-frame after final frame
in BERT mode as well as stream mode.

On receive from a file, notice EOF right away and stop
processing cleanly.

On receiving a frame with the EOS flag set, transition
to FIRST_SYNC state instead of STREAM_SYNC state. This
will still succeed (untested) if the EOS frame is immediately
followed by a frame with the same timing -- the beginning
of a new stream. But in the common case where no new stream
is immediately following, it will fail faster and go back to
searching for preamble to pick up the next transmission.