Fix overflow wraparound in polynomial interleaver
The squared term of the index computation gets a bit bigger
than 2^32 with Opulent Voice parameters,  so on 32-bit platforms
the interleaver was wrapping around and corrupting 19 bytes of
the frame data.

This fix is to use 64-bit arithmetic on all platforms. The CPU
performance hit has not been evaluated.