Get Opus voice working with protocol layers
Previous understanding of how to create an Opus packet
out of two Opus frames was wrong. The individual Opus
frames returned by opus_encode are not actually bare
frames. They include the TOC (table of contents) byte
at the beginning, so they are in fact full-fledged Opus
packets, in the configuration for a single frame per
packet. When two of them are combined into a single
packet, their individual TOC bytes need to be stripped
off and a single revised TOC byte added at the beginning
of the new packet. This is now done correctly, using the
libopus repacketizer API rather than coding our own.

Previous calculation of the COBS encoding overhead was
off by a byte. In addition to the COBS overhead itself,
which is exactly one byte for any packet of 254 bytes or
fewer, we need an additional byte of 0 to indicate the
boundary between packets. Two bytes total overhead,
rather than the one byte we had allowed.

Luckily, correct handling of the Opus packet uses two
fewer bytes per frame, so we actually come out with one
spare byte. Rather than redo the numerology again, this
version adds an extra 0 byte at the end of the COBS
encoded payload. This is automatically ignored by the
COBS decoder on receive.

UDP and IP headers are dummied out more realistically now.
The RTP header is still just a recognizable string rather
than plausible header data.

opv-mod and opv-demod both return the libopus version info
in addition to their own version numbers when invoked
with -V.

viterbi_cost is no longer used to blank packets going to
the Opus decoder. Instead, we rely on the error handling
within Opus to do something reasonable. Or, if invoked with
-b or --noise-blanker, we output silence for that frame and
don't tell the Opus decoder at all. In the future, we will
need to do something smarter to handle cases where OPV frames
arrive damaged and don't come out of the COBS decoder intact.