Remove padding byte from overhead
We had a padding byte to avoid changing the interleaver size when we added Opus re-packetization to properly combine two 20ms Opus frames. Re-packetization eliminates the TOC byte in each combined frame and adds one TOC byte for the whole combined frame, which saves a net of one byte when combining two 20ms Opus frames. But now, we're just using a single 40ms frame, which also reduces the TOC bytes from two to one, but doesn't waste the saved byte.

This is a BREAKING CHANGE because it changes the symbol rate from 27300 to 27100, but it still does not change the interleaver.